requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pymongo>=3.12.0
redis>=4.0.0
neo4j>=4.4.0
python-dotenv>=0.19.0
pydantic>=1.8.2
httpx>=0.19.0
qdrant-client>=0.7.0
openai>=0.27.0
networkx>=2.5
matplotlib>=3.4
requests>=2.25.1
schedule>=1.2.0
pyyaml>=5.4.1